cap needs to be an int.  Solving the problem by adding a cast.

+++ clients/src/mapiclient/dump.c
@@ -293,8 +293,8 @@ describe_table(Mapi mid, char *schema, char *tname, stream *toConsole, int forei
 	char *view = NULL;
 	size_t maxquerylen;
 	char *sname = NULL;
-	ssize_t cap;
-#define CAP(X) ((cap = (X)) < 0 ? 0 : cap)
+	int cap;
+#define CAP(X) ((cap = (int) (X)) < 0 ? 0 : cap)
 
 
 	if (schema == NULL) {
